[diffusion] organize runtime weight source resolution
