+++ python/sglang/multimodal_gen/runtime/weights/__init__.py
@@ -1,0 +1,1 @@
+"""Weight source contracts shared by runtime loaders."""
