+++ python/sglang/multimodal_gen/runtime/weights/source.py
